+++ src/main/java/seedu/address/logic/Logic.java
@@ -5,7 +5,7 @@ import java.nio.file.Path;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.logic.util.exceptions.ParseException;
 import seedu.address.model.person.Person;
 import seedu.address.storage.exceptions.StorageException;
