+++ src/main/java/seedu/address/logic/util/AddCommandParser.java
@@ -1,16 +1,16 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.util;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_ADDRESS;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
+import static seedu.address.logic.util.CliSyntax.PREFIX_ADDRESS;
+import static seedu.address.logic.util.CliSyntax.PREFIX_EMAIL;
+import static seedu.address.logic.util.CliSyntax.PREFIX_NAME;
+import static seedu.address.logic.util.CliSyntax.PREFIX_PHONE;
+import static seedu.address.logic.util.CliSyntax.PREFIX_TAG;
 
 import java.util.stream.Stream;
 
 import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.logic.util.exceptions.ParseException;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.fields.Address;
 import seedu.address.model.person.fields.Email;
