+++ src/main/java/seedu/address/logic/util/Parser.java
@@ -1,7 +1,7 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.util;
 
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.logic.util.exceptions.ParseException;
 
 /**
  * Represents a Parser that is able to parse user input into a {@code Command} of type {@code T}.
