+++ src/main/java/seedu/address/logic/util/ParserUtil.java
@@ -1,8 +1,8 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.util;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
-import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.logic.util.exceptions.ParseException;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
