+++ src/main/java/seedu/address/model/person/fields/Tags.java
@@ -9,7 +9,7 @@ import java.util.stream.Stream;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 
-import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.logic.util.exceptions.ParseException;
 import seedu.address.model.tag.Tag;
 
 /**
