+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.io.TempDir;
 
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.logic.util.exceptions.ParseException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
