+++ src/test/java/seedu/address/logic/util/FindCommandParserTest.java
@@ -1,8 +1,8 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.util;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
-import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
+import static seedu.address.logic.util.CommandParserTestUtil.assertParseFailure;
+import static seedu.address.logic.util.CommandParserTestUtil.assertParseSuccess;
 
 import java.util.Arrays;
