+++ src/test/java/seedu/address/model/person/fields/TagsTest.java
@@ -9,7 +9,7 @@ import java.util.Collections;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.logic.util.exceptions.ParseException;
 import seedu.address.model.tag.Tag;
 
 class TagsTest {
