+++ src/test/java/seedu/address/testutil/PersonUtil.java
@@ -1,10 +1,10 @@
 package seedu.address.testutil;
 
-import static seedu.address.logic.parser.CliSyntax.PREFIX_ADDRESS;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
+import static seedu.address.logic.util.CliSyntax.PREFIX_ADDRESS;
+import static seedu.address.logic.util.CliSyntax.PREFIX_EMAIL;
+import static seedu.address.logic.util.CliSyntax.PREFIX_NAME;
+import static seedu.address.logic.util.CliSyntax.PREFIX_PHONE;
+import static seedu.address.logic.util.CliSyntax.PREFIX_TAG;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.EditCommand;
